pyproject.toml
[tool.poetry]
name = "copilot-python"
version = "0.1.0"
description = ""
authors = ["Faisal Fida <69955157+faisal-fida@users.noreply.github.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.0"
httpx = "^0.27.0"
uvicorn = "^0.29.0"
python-dotenv = "^1.0.1"
ruff = "^0.3.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"